sim/requirements.txt
requests==2.32.3
numpy==2.1.0
python-dotenv==1.0.1
